Update: CommonInput 디자인 수정
