+++ app-frontend/src/common/CommonSelect.js
@@ -1,100 +1,132 @@
 import React, { useState } from "react";
+import Box from "@mui/material/Box";
+import Select from "@mui/material/Select";
+import MenuItem from "@mui/material/MenuItem";
+import FormControl from "@mui/material/FormControl";
 import styled from "styled-components";
 
-const SelectStyle = styled.select`
-  width: ${(props) => (props.width ? props.width : "211px")};
-  height: ${(props) => (props.height ? props.height : "38px")};
-  padding: 0 10px;
-  background-color: ${(props) =>
-    props.background ? props.background : "#F8F9FE"};
-  border: ${(props) =>
-    props.borderColor ? `1px solid ${props.borderColor}` : "1px solid #D7D7D7"};
-  border-radius: ${(props) =>
-    props.borderRadius ? props.borderRadius : "5px"};
-  font-size: ${(props) => (props.fontSize ? props.fontSize : "16px")};
+const StyledSelect = styled(Select)`
+  width: ${(props) => props.width || "211px"};
+  height: ${(props) => props.height || "38px"};
+  background-color: ${(props) => props.background || "#F8F9FE"};
+  color: ${(props) => props.color || "#b0b0b0"} !important;
+  border-radius: ${(props) => props.borderRadius || "5px"} !important;
+  font-size: ${(props) => props.fontSize || "16px"} !important;
 
-  cursor: pointer;
-  &:hover {
-    background-color: ${(props) =>
-      props.hoverBk ? props.hoverBk : props.theme.color.lightGrey};
-    color: ${(props) => (props.hoverColor ? props.hoverColor : "#757575")};
-  }
-  &:disabled {
-    background-color: #f2f2f2;
-    cursor: not-allowed;
+  &.MuiOutlinedInput-root {
+    & fieldset {
+      border-color: ${(props) => props.fieldBorderColor || "#D7D7D7"};
+    }
+
+    &:hover fieldset {
+      border-color: ${(props) => props.fieldHoverBorderColor || "#D7D7D7"};
+    }
   }
 `;
 
 const TextStyle = styled.p`
-  font-size: ${(props) => (props.fontSize ? props.fontSize : "16px")};
-  color: ${(props) => (props.color ? props.color : "#4A4A4A")};
+  font-size: ${(props) =>
+    props.fontSize ? props.fontSize : "16px"};
+  color: ${(props) => (props.color ? props.color : props.theme.color.white)};
   margin-bottom: 6px;
   margin-top: 0px;
 `;
 
-const CustomSelect = (props) => {
-  const {
-    options,
-    width,
-    height,
-    background,
-    color,
-    borderColor,
-    borderRadius,
-    fontSize,
-    disabled,
-    onChange,
-    hoverBk,
-    hoverColor,
-    text,
-  } = props;
 
+const CommonSelect = ({
+  text,
+  height,
+  options = [],
+  width,
+  background,
+  color,
+  labelColor,
+  borderRadius,
+  hoverBackground,
+  fontSize,
+  fieldBorderColor,
+  fieldHoverBorderColor,
+  fieldFocusedBorderColor,
+  onSelect,
+}) => {
   const [selectedValue, setSelectedValue] = useState("");
 
   const handleChange = (event) => {
     setSelectedValue(event.target.value);
-    if (onChange) onChange(event);
+    if (onSelect) onSelect(event.target.value);
   };
 
   return (
-    <div>
-      <TextStyle color={color} fontSize={fontSize}>
+    <Box 
+      sx={{ 
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+      }}
+    >
+      <FormControl variant="outlined">
+      <TextStyle color={labelColor} fontSize={fontSize}>
         {text}
       </TextStyle>
 
-      <SelectStyle
-        width={width}
-        height={height}
-        background={background}
-        color={color}
-        borderColor={borderColor}
-        borderRadius={borderRadius}
-        hoverBk={hoverBk}
-        hoverColor={hoverColor}
-        fontSize={fontSize}
-        disabled={disabled}
-        onChange={handleChange}
-        value={selectedValue}
-      >
-        {/* TODO: default 값으로 (맨 처음 값으로 표기) 설정 */}
-        <option disabled>
-          카드를 선택하세요 
-        </option>
-        {options.map((option, index) => (
-          <option
-            key={index}
-            value={option.value}
-            // TODO: input option text 색상 #757575 설정
-            style={{
-              fontSize: fontSize || "16px",
-            }}
-          >
-            {option.label}
-          </option>
-        ))}
-      </SelectStyle>
-    </div>
+        <StyledSelect
+          value={selectedValue === "" ? "default" : selectedValue}
+          onChange={handleChange}
+          width={width}
+          height={height}
+          background={background}
+          fontSize={fontSize}
+          fieldBorderColor={fieldBorderColor}
+          fieldHoverBorderColor={fieldHoverBorderColor}
+          fieldFocusedBorderColor={fieldFocusedBorderColor}
+          color={color}
+          borderRadius={borderRadius}
+          hoverBackground={hoverBackground}
+        >
+          {selectedValue === "" && (
+            <MenuItem value="default" disabled>
+              카드를 선택하세요
+            </MenuItem>
+          )}
+          {options.map((option, index) => (
+            <MenuItem 
+              key={index} 
+              value={option.value} 
+              sx={{ fontSize: fontSize,
+              }}>
+              {option.label}
+            </MenuItem>
+          ))}
+        </StyledSelect>
+      </FormControl>
+    </Box>
   );
 };
 
-export default CustomSelect;
+export default CommonSelect;
+
+
+/*
+사용 예시)
+
+const options = [
+  { value: "option1", label: "Option 1" },
+  { value: "option2", label: "Option 2" },
+  { value: "option3", label: "Option 3" },
+];
+
+const handleSelect = (selectedValue) => {
+  alert(`Selected value: ${selectedValue}`);
+};
+
+
+
+return 사용 예시)
+<CommonSelect
+  // options, onSelect 는 필수
+  options={options}
+  onSelect={handleSelect}
+
+  // 외에 다른 props 는 설정하지 않을시 default
+/>
+*/
